fix(server):fix all warning

+++ game-lib/include/powerup/PowerupRegistry.hpp
@@ -34,14 +34,14 @@ struct PowerupDefinition {
     PowerupCategory category;
     uint8_t max_level;
     std::string asset_path;
-    
+
     struct LevelEffect {
         float duration;
         float value;
         float cooldown;
         std::string description;
     };
-    
+
     std::vector<LevelEffect> level_effects;
 };
 
@@ -53,28 +53,28 @@ public:
     }
 
     void initialize();
-    
+
     const PowerupDefinition* get_powerup(PowerupId id) const;
     const PowerupDefinition* get_powerup(uint8_t id) const;
-    
+
     std::vector<PowerupId> get_all_powerup_ids() const;
     std::vector<PowerupId> get_powerups_by_category(PowerupCategory category) const;
-    
+
     PowerupId get_upgraded_version(PowerupId id, uint8_t current_level) const;
-    
+
     PowerupRegistry(const PowerupRegistry&) = delete;
     PowerupRegistry& operator=(const PowerupRegistry&) = delete;
 
 private:
     PowerupRegistry() = default;
-    
+
     std::unordered_map<PowerupId, PowerupDefinition> powerups_;
-    
+
     void register_powerup(const PowerupDefinition& def);
 };
 
 inline uint16_t make_powerup_key(PowerupId id, uint8_t level) {
-    return (static_cast<uint16_t>(id) << 8) | level;
+    return static_cast<uint16_t>((static_cast<uint16_t>(id) << 8) | static_cast<uint16_t>(level));
 }
 
 inline PowerupId extract_powerup_id(uint16_t key) {
